% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fftfreq_kim.R
\name{fftfreq_kim}
\alias{fftfreq_kim}
\title{Computes k and l for FFT}
\usage{
fftfreq_kim(dx, dy, nxpad, nypad)
}
\arguments{
\item{dx}{resolution in x direction}

\item{dy}{resolution in y direction}

\item{nxpad}{number of columns in dem}

\item{nypad}{number of rows in in dem}
}
\description{
Computes k and l for FFT
}
\details{
as the name suggests, this is Kim's approach. Don't use this because it doesn't seem to replicate the guassian hill example.
}
