% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fftfreq_qgis.R
\name{fftfreq_qgis}
\alias{fftfreq_qgis}
\title{Computes k and l matrixes for FFT}
\usage{
fftfreq_qgis(dx, dy, nxpad, nypad)
}
\arguments{
\item{dx}{distance per grid cell in x direction}

\item{dy}{distance per grid cell in y direction}

\item{nxpad}{number of columns}

\item{nypad}{number of rows}
}
\description{
Computes k and l matrixes for FFT
}
