% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_dem.R
\name{gaussian_dem}
\alias{gaussian_dem}
\title{Replicates the topography used in Figures 3 and 4 in Smith and Barstad 2004}
\usage{
gaussian_dem(dx, dy)
}
\arguments{
\item{dx}{resolution along x axis}

\item{dy}{resolution along y axis}
}
\description{
Replicates the topography used in Figures 3 and 4 in Smith and Barstad 2004
}
