% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_matrix.R
\name{image_matrix}
\alias{image_matrix}
\title{A function that prints matrices properly, with legend.}
\usage{
image_matrix(x, ...)
}
\arguments{
\item{x}{matrix to plot}
}
\description{
A function that prints matrices properly, with legend.
}
\details{
credit to http://www.phaget4.org/R/image_matrix.html
}
