% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertical_wavenumber.R
\name{vertical_wavenumber}
\alias{vertical_wavenumber}
\title{Computes the vertical wavenumber}
\usage{
vertical_wavenumber(ny, k, l, method = NULL, coriollis = FALSE)
}
\arguments{
\item{ny}{if ny=1 wavenumber is for a 1D profile}

\item{k}{sampling freq for x}

\item{l}{sampling freq for y}

\item{method}{default is kim's approach. 'qgis' for qgis approach}
}
\description{
Computes the vertical wavenumber
}
\details{
the default method is Kim's
}
