% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTmodel.R
\name{LTmodel}
\alias{LTmodel}
\title{Run the Linear Theory model. THis is the function that calls the others.}
\usage{
LTmodel(h, dx, dy, Pinf, tauc, tauf, windspeed, winddir, trunc = TRUE,
  method = "smith", trunc_value = 0, saveInt = FALSE, Intfn = NULL)
}
\arguments{
\item{h}{matrix of elevation}

\item{dx}{horizontal resolution (m)}

\item{dy}{vertical resolution (m)}

\item{Pinf}{background precip (mm/hr)}

\item{tauc}{condensation time delay (s)}

\item{tauf}{fallout time delay (s)}

\item{windspeed}{windspeed (m/s)}

\item{winddir}{wind direction (m/s)}

\item{trunc}{should the result truncate negative values (default=TRUE)}

\item{method}{default = 'smith' (as the method is described in Smith and Barstad 2004); decides whether truncation should happen before or after adding background precip. If 'smith', truncation happens after adding background precip. If otherwise, truncation happens before adding background precip. For 'smith', you will have a lot of zeros if background precip is low whereas with the other method those grid cells will have the value of the background precip}

\item{trunc_value}{value below which modeled precip rates should go to zero. This will only make a difference if `method=LEM`}

\item{saveInt}{logical. save fft of dem to 'int' directory of your project. speeds up parameter search 2x-3x}

\item{Intfn}{filename (relative to wd) where to save intermediate FFT matrix}
}
\description{
Run the Linear Theory model. THis is the function that calls the others.
}
\details{
returns a matrix the same size as h with precip mm/hr
Prepare Fourier transform of the topography ----
Do 2D FFT of Topography ----
Prepare precipitation transfer function relating the Fourier transforms of the terrain hhat(k,l) and the precipitiation field Phat(k,l) (Eq 49)
Transfer function relating the Fourier transforms of the terrain hhat(k,l) and the precipitiation field Phat(k,l) (Eq 49)
Perform the inverse Fourier transform on Phat
Should background precip be added before or after truncation (if truncation happens)?
Add the background precipitation rate
Apply the positive cutoff
Apply the positive cutoff
Add the background precipitation rate
library(tidyverse)
tbl <- tibble(x=as.numeric(dem_list$x),y=as.numeric(dem_list$y),p=Ppad)
Return to matrix
}
